ghc: Drop broken sphinx HTML docs gen
Problem:

 - FTBFS w/sphinx >=v6.1.3
 - Tested w/v7.0.1

Summary:

FTBFS with sphinx >=6.1.3 when generating HTML docs (the default).
Upstream says it's a known issue, requiring backport of patches and
supposedly fixed in 9.6.1, but testing reveals it still fails w/sphinx v7.0.1

History of this formula applied patch for ghc 9.4.4, sphinx-doc 6+, but this
still failed to build with or without that patch.
Homebrew maintainer removed this patch in 23262c32bc3 PR #136863

Suggest that we disable HTML doc generation to avoid this & future FTBFS issues.

Upstream issues:

  - https://gitlab.haskell.org/ghc/ghc/-/issues/23286
  - https://gitlab.haskell.org/ghc/ghc/-/issues/23477